Refactor filesystem path/pack/bootstrap logic into core modules
