+++ Quake/Makefile.w32
@@ -272,6 +272,9 @@ OBJS = strlcat.o \
         wad.o \
         cmd.o \
         common.o \
+        fs_searchpath.o \
+        fs_pack.o \
+        bootstrap.o \
 	mat_material.o \
 	mat_material_parse.o \
         steam.o \
