+++ Quake/Makefile.w64
@@ -265,6 +265,9 @@ OBJS = strlcat.o \
         wad.o \
         cmd.o \
         common.o \
+        fs_searchpath.o \
+        fs_pack.o \
+        bootstrap.o \
 	mat_material.o \
 	mat_material_parse.o \
         steam.o \
