+++ Quake/src/audio/snd_codec.c
@@ -299,7 +299,7 @@ snd_stream_t *S_CodecUtilOpen(const char *filename, snd_codec_t *codec, qboolean
 
 	/* Try to open the file */
 	length = (long) COM_FOpenFile(filename, &handle, NULL);
-	pak = file_from_pak;
+	pak = COM_GetFileFromPak ();
 	if (length == -1)
 	{
 		Con_DPrintf("Couldn't open %s\n", filename);
